+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 <!-- next-header -->
 
 ## [Unreleased] - ReleaseDate
+
+## [0.9.0] - 2021-12-01
 ### Changed
 - UI change: Show background, border, and aux as separate color patches.
 This makes it impossible to try to draw with colors 8-15 unless in a multicolor cell, and then only using aux or bg.
@@ -54,7 +56,8 @@ This makes it impossible to try to draw with colors 8-15 unless in a multicolor 
 
 
 <!-- next-url -->
-[Unreleased]: https://github.com/vilcans/pixel_pen/compare/v0.8.0...HEAD
+[Unreleased]: https://github.com/vilcans/pixel_pen/compare/v0.9.0...HEAD
+[0.9.0]: https://github.com/vilcans/pixel_pen/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/vilcans/pixel_pen/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/vilcans/pixel_pen/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/vilcans/pixel_pen/compare/v0.5.0...v0.6.0